+++ lib/rack-openid2.rb
@@ -2,8 +2,8 @@
 require "version_gem"
 require_relative "rack/openid/version"
 
-require_relative "rack/openid"
-
-Rack::Version.class_eval do
+Rack::OpenID::Version.class_eval do
   extend VersionGem::Basic
 end
+
+require_relative "rack/openid"